Moved business logic into ApplicationService
