+++ src/Backend/Models/Response/SubscriptionRegistrationRequest.cs
@@ -1,8 +1,7 @@
 ﻿namespace Backend.Models.Response;
 
-using System;
-
 public record SubscriptionRegistrationOkResponse(string ClientId,  string ClientSecret, string TenantID);
 
 public record SubscriptionRegistrationFailedResponse(string Message);
 
+public record CreateServicePrincipalInKeyVaultResponse(string ClientId, string SecretURL, string TenantID);
